requirements.txt
google-generativeai==0.3.1
IPython
textwrap
streamlit